requirements.txt
annoy==1.17.3
asttokens==2.4.0
backcall==0.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
dataclasses-json==0.6.1
decorator==5.1.1
diskcache==5.6.3
executing==2.0.0
faiss-cpu==1.7.4
fasteners==0.19
fasttext==0.9.2
ffmpeg-python==0.2.0
filelock==3.12.4
fsspec==2023.9.2
future==0.18.3
hnswlib==0.7.0
huggingface-hub==0.17.3
idna==3.4
ipyplot==1.1.1
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
llama_cpp_python==0.2.11
lz4==4.3.2
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numpy==1.26.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pybind11==2.11.1
Pygments==2.16.1
pymagnitude-lite==0.1.143
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
sexpdata==1.0.1
shortuuid==1.0.11
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.14.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
traitlets==5.11.2
transformers==4.34.0
triton==2.1.0
txtai @ git+https://github.com/neuml/txtai@4447685ef933aba260ca3aa4868c62321beb5cf5
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.6
Wand==0.6.11
wcwidth==0.2.8
xxhash==3.4.1
